+++ package.json
@@ -15,8 +15,10 @@
     "vite": "^3.2.3"
   },
   "dependencies": {
+    "axios": "^1.2.0",
     "leader-line-new": "^1.1.9",
     "node-sass": "^8.0.0",
-    "svelte-preprocess": "^4.10.7"
+    "svelte-preprocess": "^4.10.7",
+    "svelte-spa-router": "^3.3.0"
   }
 }
