+++ src/lib/layouts/TopBar.svelte
@@ -1,1 +1,3 @@
+<a href="/#/notfound">NotFound</a>
+<a href="/#/dragdrop">DragDrop</a>
 😎 VANETLAT FRONTEND 😎
