+++ src/lib/scenario/Scenario.svelte
@@ -3,8 +3,6 @@
     import Simulation from "./Simulation.svelte"
     import Topology from "./Topology.svelte"
 
-    import { topology, containers, apps } from '../../store/store.js';
-
     let isTopologyVisible = false
     let TopologyBntText = "Show topology settings"
     let isSimulationVisible = false
@@ -21,30 +19,6 @@
         SimulationBntText = isSimulationVisible ? "Hide simulation settings" : "Show simulation settings"
     }
 
-    function print(){
-        let top = $topology
-        let cons = $containers
-        let simu = $apps
-        top["container_setings"] = {}
-        cons.forEach(element => {
-            top["container_setings"][element.name] = element
-        })
-        let sim = {
-            "server": {},
-            "client": {}
-        }
-        simu.server.forEach(e => {
-           sim.server[e.name] = e
-        })
-        simu.client.forEach(e => {
-           sim.client[e.name] = e
-        })
-        let data = {
-            "topology": top,
-            "simulation": sim
-        }
-        console.log(JSON.stringify(data))
-    }
 </script>
 
 <div>
@@ -69,9 +43,6 @@
     </div>
     {/if}
 
-    <div>
-        <button on:click={print}>Print scenario</button>
-    </div>
 </div>
 
 <style>
