+++ src/services/Api.js
@@ -1,0 +1,11 @@
+import axios from 'axios'
+
+let api = axios.create({
+    baseURL: `http://127.0.0.1:3333`
+  })
+
+export default {
+  post(path, data){
+    return api.post(path, data)
+  }
+}
